Move code into separate components, add lint and proptypes
