+++ package.json
@@ -12,6 +12,7 @@
     "@babel/core": "^7.15.5",
     "formidable": "^1.2.2",
     "next": "^11.1.2",
+    "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
@@ -19,6 +20,9 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.1.0",
     "babel-jest": "^27.1.1",
+    "eslint": "^7.32.0",
+    "eslint-plugin-jest": "^24.4.0",
+    "eslint-plugin-react": "^7.25.1",
     "jest": "^27.1.1"
   }
 }
